Minor code adjustments

+++ Neos.Flow/Tests/Behavior/Features/Bootstrap/SecurityOperationsTrait.php
@@ -104,7 +104,7 @@ trait SecurityOperationsTrait
      */
     public static function cleanUpSecurity()
     {
-        if (file_exists(self::$testingPolicyPathAndFilename)) {
+        if (self::$testingPolicyPathAndFilename && file_exists(self::$testingPolicyPathAndFilename)) {
             unlink(self::$testingPolicyPathAndFilename);
         }
     }
